% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addBreakColorLegend.R
\name{addBreakColorLegend}
\alias{addBreakColorLegend}
\title{Adds a Legend to sp Plots with Colors and Breaks}
\usage{
addBreakColorLegend(xleft, ybottom, xright, ytop, breaks, colors,
  printFormat = "4.2", ...)
}
\arguments{
\item{xleft}{a scalar of the left x position.}

\item{ybottom}{a scalar of bottom y position.}

\item{xright}{a scalar of the right x position.}

\item{ytop}{a scalar of top y position.}

\item{colors}{a vector of colors, should have one less element than the number of breaks.}

\item{brks}{the breaks between color classes.  Needs a lower bound, and upper bound, so there should be one more break than number of classes.}

\item{printFormat:}{a character variable in '4.2' format where the 4 control the number of digits before the decimal, and 2 controls the number of digits after the decimal.}
}
\value{
add a color ramp legend as a rectangle to the currently active plot
}
\description{
Adds a Legend to sp Plots with Colors and Breaks
}
\author{
Jay Ver Hoef
}
\seealso{
\code{\link{plotPointsRGB}}, \code{\link{rect}},
}

