% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lmspar.R
\name{lmspar}
\alias{lmspar}
\title{fit a linear model with spatial CAR and SAR covariances using maximum likelihood}
\usage{
lmspar(formula, data, indComp = FALSE, Nmat = NULL, distMat = NULL,
  crsStk = NULL, indSamp, model = NULL, rowStand = TRUE,
  rhoBound = c(-1, 1))
}
\arguments{
\item{formula}{an R formula}

\item{data}{a data.frame containing the data}

\item{indComp}{should an independent component be added to the model}

\item{Nmat}{neighborhood matrix of binary values, with a 1 indicating a neighbor, 0 otherwise}

\item{distMat}{Euclidean distance matrix among polygon centroids}

\item{crsStk}{indicator matrix of whether locations cross stock boundaries}

\item{indSamp}{indicator vector for wheter location was sampled. Zero indicates missing value}

\item{model}{either 'CAR' or 'SAR'.  Only used if one of Nmat, distMat, or crsStk are not NULL.}

\item{rowStand}{should row-standardization be used on CAR or SAR models?  Default is TRUE.}
}
\value{
a list with parmEst as the estimated covariance parameters (with overall variance profiled out), sigma2 as the estimate of the overall variance parameter, m2LL as the minimized negative 2 times the log-likelihood, bHat as the fitted fixed effect parameters, covb as the estimated covariance matrix of the fitted fixed effect parameters, Vi is the inverse of the covariance matrix, Vi.oo as the inverse of the covariance matrix of the observed locations only, X returns the design matrix, and y returns the data vector.
}
\description{
fit a linear model with spatial CAR and SAR covariances using maximum likelihood
}
\author{
Jay Ver Hoef
}

