% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/m2LL.R
\name{m2LL}
\alias{m2LL}
\title{two times the negative log-likelihood}
\usage{
m2LL(theta, X, y, indComp, Nmat = NULL, distMat = NULL, crsStk = NULL,
  indSamp, model = NULL, rowStand = NULL, rhoBound = c(-1, 1))
}
\arguments{
\item{theta}{covariance parameters, with overall variance parameter profiled out.}

\item{X}{design matrix for fixed effects}

\item{y}{vector of data for response variable}

\item{indComp}{an additive independent component to the model.  Default is TRUE.}

\item{Nmat}{neighborhood matrix}

\item{distMat}{distance matrix}

\item{crsStk}{indicator matrix of whether locations cross stock boundaries}

\item{indSamp}{indicator vector for wheter location was sampled. Zero, or FALSE, indicates missing value}

\item{model}{either 'CAR' or 'SAR' to be the model associated with the Nmat argument}

\item{logical}{value on whether Nmat should be row-standardized}

\item{rhobound}{upper bound for rho.  This should be determined from the eigenvalues of Nmat prior to running function}
}
\value{
two times the negative log-likelihood
}
\description{
two times the negative log-likelihood
}
\author{
Jay Ver Hoef jverhoef
}

