% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Neighmat.r
\name{Neighmat}
\alias{Neighmat}
\title{Create Neighborhood Matrix from Adjacency List}
\usage{
Neighmat(adj, num, n)
}
\arguments{
\item{adj}{adjacency list, as created by poly2nb fron spdep package}

\item{num}{the number of adjacent neighbors for each list item above}

\item{n}{total number of list items}
}
\value{
a matrix, with 1's indicating neighbors, and 0 otherwise, in the same order as adj.
}
\description{
Create Neighborhood Matrix from Adjacency List
}
\author{
Jay Ver Hoef
}

